Adjusting buffer

+++ pru0adc.c
@@ -15,7 +15,6 @@
 #include <pru_virtqueue.h>
 #include <pru_rpmsg.h>
 #include "resource_table_0.h"
-//#include "rpmsg_send.h"
 
 // Define remoteproc related variables.
 #define HOST_INT ((uint32_t) 1 << 30)
@@ -37,7 +36,7 @@
 #define VIRTIO_CONFIG_S_DRIVER_OK  4
 
 //  Buffer used for PRU to ARM communication.
-uint8_t payload[RPMSG_BUF_SIZE];
+uint16_t payload[RPMSG_BUF_SIZE];
 
 #define PRU_SHAREDMEM 0x00010000
   volatile register uint32_t __R30;
@@ -76,7 +75,7 @@ while(1) {
          CT_INTC.SICR_bit.STS_CLR_IDX = FROM_ARM_HOST;
          //  Receive all available messages, multiple messages can be sent per kick.
         while (pru_rpmsg_receive(&transport, &src, &dst, payload, &len) == PRU_RPMSG_SUCCESS) {
-    uint8_t payloadOut[1] = {1};
+    uint16_t payloadOut[1] = {1};
     len = 1;  //  This is the length of payloadOut in bytes.
         //  Echo the message back to the same address from which we just received.
                pru_rpmsg_send(&transport, dst, src, payloadOut, len);
